path: preserve SSH prefixes when resolving relative paths
The function `git_path_resolve_relative` is not only used for paths, but
also for URLs. As a result, it needs to detect certain URL prefixes like
for example schemata ("http://") which are not to be removed during any
directory traversals.

We already handle the schemata part correctly, but we are missing the
special case of SSH. In contrast to most other protocols, SSH uses a
colon instead of a slash after the host part, like for example in
"git@example.com:foo". As a result, `git_path_resolve_relative` would've
treated the complete "git@example.com:foo" part as a single directory
part and thus remove it during upwards-traversal.

This problem is relatively easy to hit when having submodules with a
relative URL. As `git_submodule_resolve_url` calls out to
`git_path_resolve_relative`, it mis-treated SSH URLs and thus didn't
compute them correctly.

Fix the issue by special-casing the SSH protocol. Detection makes use
either of the "ssh://" prefix or with a heuristic that tries to detect
whether there's a colon before a slash. Note that especially the second
heuristic is error prone, but it is in fact what git uses right now.

Add some tests to verify behaviour.