worktree: add flag to allow re-using existing branches for new worktrees
In this commit's parent, we have introduced logic that will
automatically re-use of existing branches if the new worktree name
matches the branch name. While this is a handy feature, it changes
behaviour in a backwards-incompatible way and might thus surprise users.
Furthermore, it's impossible to tell whether we have created the
worktree with a new or an existing reference.

To fix this, introduce a new option `checkout_existing` that toggles
this behaviour. Only if the flag is set will we now allow re-use of
existing branches, while it's set to "off" by default.