editorconfig: special-case Python scripts
Python's PEP 8 specifies that one shall use spaces instead of tabs as
coding style, and we actually honor that currently. Our EditorConfig
does not special-case Python scripts, though, which is why we end up
with our C coding style and thus with tabs.

Special-case "*.py" files to override that default with spaces to fix
this.