win32: fix path canonicalization for paths starting with "../"
The path canonicalization function for Windows strips leading "../".
This behaviour has been introduced with commit cceae9a25 (win32: use
NT-prefixed "\\?\" paths, 2014-12-01), saying that

> we must not rely on the system to translate "." and ".." in paths

The reasoning is that we mostly try to use NT-prefixed "\\?\" paths,
which does turn off path canonicalization. What's weird though is that
if we simply strip leading "..", then there is no way that we can end up
with the correct path, as we do not adjust any of the following path
components accordingly.

One place where this bites us is with the path canonicalization
for symlink targets introduced with commit 7d55bee6d (win32: fix
relative symlinks pointing into dirs, 2020-01-10). If creating a symlink
with target "../target", then the canonicalization function will strip
the leading "../" and create a symlink with target "target".

Given that stripping this prefix cannot be correct in general, let's
remove the stripping code and instead leave such prefixes intact. If
understanding the code correctly, then the change _should_ fix cases as
with the symlink while not breaking anything else.