azure: upgrade to newer Windows VM images
This is a subset of commit 95f329b49 (azure: upgrade to newer hosted VM
images, 2020-03-10), upgrading all of our Windows jobs to use
'vs2017-win2016' machines and macOS to 'macos-10.14'. This is intended
to keep our continuous integration builds from failing in the future, as
these images will get deprecated on March 31st. As this is in
preparation of a stable release, we do not want to upgrade any of the
other machines like is done in the mentioned commit but keep the impact
minimal.

fixup! azure: upgrade to newer Windows VM images