errors: make git_error_set public
Make `git_error_set` public so that the forthcoming utility library can
take advantage of it.  Although this is not dangerous to export, it is
not up to our general standards of public API since it cannot return an
error code, for instance.  Therefore, we discourage its use in a comment
and by hiding it in a `sys` include directory.