repository: use state machine to determine repo state
The current logic to determine a repo's state is really repetetive and
thus both hard to read and easy to make mistakes. Furthermore, we need
to extend it to look up references via the refdb and not via the
filesystem.

So as a prepratory step, this commit refactors the function to use a
state machine of states to avoid repetiton and make the logic easier to
extend.