repository: determine state via refdb
In order to determine a repo's state we need to look at a set of
references, amongst other things. This check is currently done by
checking whether the reference exists on-disk directly and thus
circumvents the refdb abstraction.

Fix this by introducing a new `COND_REF` condition that uses
`git_refdb_exists` instead of a filesystem-based check.