Implement generic TLS interface
This adds a generic TLS interface for anyone to store TLS data. It is designed to work regardless of whether threading support is built into the library or not.

Nobody in the library should directly interface with the data on the TLS struct, so it's been built to be opaque even in the library.

Requires the allocator to be initialized before use.