Introduce clang-tidy
This change adds -DENABLE_CLANG_TIDY, which makes the compilation invoke
clang-tidy in parallel. This should help catch a lot more bugs. It also
enables it for the UBSan CI job to avoid slowing the whole CI down.

As proof of concept, only the clang-analyzer-core.NullDereference and a
few other core checks were enabled to avoid changing a ton of things at
once.

A potential memory leak in `annotated_commit_init()` was found and fixed
with this.