Refactor clar__fail()
This change makes clar__fail() always abort the test. The
__attribute__((__noreturn__)) is added to clar__fail() so that the
static analyzers can reason about this fact, and clar__assert() is
rewritten as a macro so that clar__fail() is always called inline.
Since there is now a need to have a way to fail a test without abortting
it, clar__warn() is introduced for that purpose.